Extract HistoryCallbacks
Previously named LogCallbacks
