+++ sciencer/utils/__init__.py
@@ -1,3 +1,4 @@
 """Utilities for Sciencer toolkit
 """
 from .csv_callback import WriteToCSVCallbacks
+from .history_callback import HistoryCallbacks
